Ensure the Action classes exist before trying to hook to them
vagrant-aws v0.4.0 will remove the custom TimedProvision class and use
only the standard Builtin::Provision instead. To support both old and
new versions of the plugin, we check if the TimedProvision class is
defined before using it.

While at it, do the same check for vagrant-omnibus plugin, just in case
it also changes in the future.