Ensure that Yum configuration includes the (default) port
Yum requires that the proxy URI includes the port number. But
`URI::Generic#to_s` drops the port if it's the default for the specified
scheme, e.g. `80` for HTTP URIs.

So make sure that the port is always included. Also drop the path
(normally just a slash) as HTTP proxy URIs really don't use use.

Thanks to @partyzan and @johnbellone for all the information and help!

Fixes #36.