Configure Proxy Before not After Omnibus
Was trying to get the vagrant-omnibus plugin working behind a corporate proxy today and it was working fine for everything else but timing out for the omnibus plugin.  I'm very new to ruby/vagrant/etc but I'm thinking this is the reason.

Thanks for all your work, this plugin has cured almost all my proxy headaches.